Some restructuring of the classes. Moved a lot of functionality into the generic Object class per Andrew's suggestion.
Objects now load their surfaces from the image specified in its constructor.
setWindow() reformats the object surface to match that of the GameWindow.

Added a data folder to store maps, images, and configuration options.
So far there's only a player1.jpg used for the player image.
It's a random image of mario i found on the internet.