Created isometric drawing engine. Mario now moves about on a grass/dirt tileset.
Allow Objects to share surfaces. This is especially useful for tilesets, where having a tileset surface for each tile is prohibitively expensive in terms of memory. Added a surfaceIsShared bool that indicates whether an object contains a shared surface or not. If it does, we do not free it upon destruction and we do not optimize it for display. Rather we let the original creator of the surface handle that

Removed SDL_Surface* from Player and Object constructors. Set surface manually after creation. This is to facilitate shared surfaces.